Split maptools
